Update README.md
A time table management system is a project which is developed to provide better support for staff in a college. The time table management system provides a function to create and view the specific staff and student's time table. This system also contains a database that stores staff and student allotted subject details and notifications before time. A college timetable is a temporal arrangement of a set of staff and students in which all given constraints are satisfied. Hence we have developed a practical approach for building staff and student timetable systems, which can be customized to fit any college timetable system. Each staff and student can view their timetable once they are finalized for a given semester and can also edit the timetable. This project utilizes HTML CSS and JavaScript to create an intuitive and userfriendly interface for Admin and faculty to save their time.